Hack-O-Hire-TechElevate-DJango/webapi/requirements.txt
Django==4.0.3
opentelemetry-sdk
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-django==0.29b0
opentelemetry-instrumentation-logging==0.29b0
opentelemetry-instrumentation-requests==0.29b0
requests==2.27.1
django-prometheus==2.2.0
mysql-connector-python
mysqlclient
structlog
python-json-logger